+++ src/test/java/leetcode/medium/SortListTest.java
@@ -56,63 +56,4 @@ class SortListTest {
       result = result.next;
     }
   }
-
-  class SortList {
-    public ListNode sort(ListNode head) {
-      if (head == null)
-        return null;
-      ListNode dummy = new ListNode(0);
-      dummy.next = head;
-      int size = 0;
-      while (head != null) {
-        head = head.next;
-        size++;
-      }
-
-      for (int step = 1; step < size; step *= 2) {
-        ListNode prev = dummy;
-        ListNode cur = dummy.next;
-        while (cur != null) {
-          ListNode left = cur;
-          ListNode right = divide(left, step);
-          cur = divide(right, step);
-          prev = merge(left, right, prev);
-        }
-      }
-      return dummy.next;
-    }
-
-    private ListNode divide(ListNode left, int step) {
-      if (left == null || left.next == null)
-        return null;
-
-      for (int i = 1; left.next != null && i < step; i++)
-        left = left.next;
-      
-      ListNode right = left.next;
-      left.next = null;
-      return right;
-    }
-
-    private ListNode merge(ListNode left, ListNode right, ListNode sorted) {
-      while (left != null && right != null) {
-        if (left.val <= right.val) {
-          sorted.next = left;
-          left = left.next;
-        } else {
-          sorted.next = right;
-          right = right.next;
-        }
-        sorted = sorted.next;
-      }
-
-      if (left == null)
-        sorted.next = right;
-      if (right == null)
-        sorted.next = left;
-      while (sorted.next != null)
-        sorted = sorted.next;
-      return sorted;
-    }
-  }
 }
